Added feature to auto rename template file.
This feature allow us to write non template file and when applied should be
changed by template file. eg: we want write `README.md` that should displayed
on github to describe about the reframe source, this README.md should not be part of
the Reframe when applied, so we can write "the real" README.md for Reframe user
with `.template` postfix followed by the extension on it file name, eg: `README.template.md`,
when this template processed by Reframe, the `README.template.md` will replace `README.md`
in place.